requirements.txt
streamlit==1.44.0
torch
torchvision
pillow